Event functionality complete
