+++ src/App.js
@@ -2,7 +2,7 @@ import { useState } from "react";
 import eventsData from "./data";
 
 // import Attendees from "./Attendees";
-// import Event from "./Components/Event";
+import Event from "./Components/Event";
 import Footer from "./Components/Footer";
 import Header from "./Components/Header";
 import NewEventForm from "./Components/NewEventForm";
@@ -10,14 +10,8 @@ import NewEventForm from "./Components/NewEventForm";
 function App() {
   const [events, setEvents] = useState(eventsData);
 
-  const [showAttendees, setShowAttendees] = useState(false);
-
   function handleAddEvent(event) {
     setEvents([event, ...events]);
-  }
-
-  function toggleEventAttendees() {
-    setShowAttendees(!showAttendees);
   }
 
   function updateEventAttendance(eventId, attendeeId) {
@@ -44,51 +38,7 @@ function App() {
             {events.map((event) => {
               const { people: attendees } = event;
 
-              return (
-                <>
-                  <li key={event.id}>
-                    <img src={event.eventImage} alt={event.name} />
-                    <h5>
-                      {event.name} {event.eventType}
-                    </h5>
-                    <br />
-                    <span>Organized by: {event.organizer} </span>
-                    <br />
-                    <>
-                      <button onClick={toggleEventAttendees}>{!showAttendees ? "Show Attendees" : "Hide Attendees"}</button>
-
-                      {showAttendees ? (
-                        <div className="attendees">
-                          {attendees.map((attendee, index) => (
-                            <>
-                              <div key={attendee.id} className="attendee">
-                                <p>
-                                  <img src={attendee.avatar} alt={attendee.firstName} />
-                                  {"   "}
-                                  <span>
-                                    {" "}
-                                    {attendee.firstName} {attendee.lastName}{" "}
-                                  </span>
-                                </p>
-                                <p>
-                                  <button className="clickable" onClick={() => updateEventAttendance(event.id, attendee.id)}>
-                                    Attending:
-                                  </button>
-                                  <span>{attendee.attendance ? "✅" : "❌"}</span>
-                                </p>
-
-                                <p>
-                                  <span>Note:</span> {attendee.note}
-                                </p>
-                              </div>
-                            </>
-                          ))}
-                        </div>
-                      ) : null}
-                    </>
-                  </li>
-                </>
-              );
+              return <Event event={event} attendees={attendees} updateEventAttendance={updateEventAttendance} />;
             })}
           </ul>
         </div>
